Add http logging middleware
