+++ vclogin/pages/api/clientMetadata.ts
@@ -6,11 +6,9 @@
 import { getMetadata } from "@/lib/getMetadata";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { logger } from "@/config/logger";
+import { withLogging } from "@/middleware/logging";
 
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<any>,
-) {
+async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
   try {
     const { method } = req;
     if (method === "GET") {
@@ -28,3 +26,4 @@ export default async function handler(
 }
 
 export const config = { api: { bodyParser: false } };
+export default withLogging(handler);
